+++ oracle/oracle.go
@@ -1,18 +1,15 @@
 package oracle
 
 import (
-	"encoding/binary"
+	"crypto/aes"
+	"crypto/cipher"
 	"errors"
 	"fmt"
-
-	"github.com/spacemeshos/sha256-simd"
 
 	"github.com/spacemeshos/post/config"
 	"github.com/spacemeshos/post/gpu"
 	"github.com/spacemeshos/post/shared"
 )
-
-type Challenge = shared.Challenge
 
 type option struct {
 	computeProviderID uint
@@ -171,12 +168,21 @@ func WorkOracle(opts ...OptionFunc) (WorkOracleResult, error) {
 	}, nil
 }
 
-func FastOracle(ch Challenge, nonce uint32, label []byte) [32]byte {
-	input := make([]byte, 32+4+len(label))
+// CreateBlockCipher creates an AES cipher for given fast oracle block.
+// A cipher is created using an idx encrypted with challenge:
+//
+//	cipher = AES(AES(ch).Encrypt(i))
+func CreateBlockCipher(ch shared.Challenge, nonce uint8) (cipher.Block, error) {
+	// A temporary cipher used only to create key.
+	// The key is a block encrypted with AES which key is the challenge.
+	keyCipher, err := aes.NewCipher(ch)
+	if err != nil {
+		return nil, err
+	}
 
-	copy(input, ch)
-	binary.LittleEndian.PutUint32(input[32:], nonce)
-	copy(input[36:], label[:])
-
-	return sha256.Sum256(input)
+	keyBuffer := make([]byte, aes.BlockSize)
+	keyBuffer[0] = nonce
+	key := make([]byte, aes.BlockSize)
+	keyCipher.Encrypt(key, keyBuffer)
+	return aes.NewCipher(key)
 }
