+++ proving/proving_options.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"io"
 
+	"github.com/spacemeshos/post/config"
 	"github.com/spacemeshos/post/initialization"
 	"github.com/spacemeshos/post/persistence"
 	"github.com/spacemeshos/post/shared"
@@ -29,7 +30,7 @@ func (o *option) validate() error {
 type OptionFunc func(*option) error
 
 // WithDataSource sets the data source to use for the proof.
-func WithDataSource(cfg Config, nodeId, commitmentAtxId []byte, datadir string) OptionFunc {
+func WithDataSource(cfg config.Config, nodeId, commitmentAtxId []byte, datadir string) OptionFunc {
 	return func(o *option) error {
 		m, err := initialization.LoadMetadata(datadir)
 		if err != nil {
