+++ verifying/verifying.go
@@ -2,23 +2,15 @@ package verifying
 
 import (
 	"bytes"
+	"crypto/aes"
+	"crypto/cipher"
 	"errors"
 	"fmt"
 	"math/big"
+	"unsafe"
 
-	"github.com/spacemeshos/post/config"
 	"github.com/spacemeshos/post/oracle"
 	"github.com/spacemeshos/post/shared"
-)
-
-type (
-	Config = config.Config
-)
-
-var (
-	WorkOracle = oracle.WorkOracle
-	FastOracle = oracle.FastOracle
-	UInt64LE   = shared.UInt64LE
 )
 
 // VerifyVRFNonce ensures the validity of a nonce for a given node.
@@ -40,7 +32,7 @@ func VerifyVRFNonce(nonce *uint64, m *shared.VRFNonceMetadata) error {
 	difficulty := shared.PowDifficulty(numLabels)
 	threshold := new(big.Int).SetBytes(difficulty)
 
-	res, err := WorkOracle(
+	res, err := oracle.WorkOracle(
 		oracle.WithCommitment(oracle.CommitmentBytes(m.NodeId, m.CommitmentAtxId)),
 		oracle.WithPosition(*nonce),
 		oracle.WithBitsPerLabel(uint32(m.BitsPerLabel)*32),
@@ -59,15 +51,25 @@ func VerifyVRFNonce(nonce *uint64, m *shared.VRFNonceMetadata) error {
 
 // Verify ensures the validity of a proof in respect to its metadata.
 // It returns nil if the proof is valid or an error describing the failure, otherwise.
-//
-// Deprecated: Verify is deprecated. Use the verifying.VerifyNew function instead.
-func Verify(p *shared.Proof, m *shared.ProofMetadata) error {
+func Verify(p *shared.Proof, m *shared.ProofMetadata, opts ...OptionFunc) error {
+	options := &option{
+		logger: &shared.DisabledLogger{},
+	}
+	for _, opt := range opts {
+		opt(options)
+	}
+	if err := options.validate(); err != nil {
+		return err
+	}
+
+	if (m.BitsPerLabel) != 8 {
+		return fmt.Errorf("invalid `bitsPerLabel, only 8-bit label is supported, given: %v", m.BitsPerLabel)
+	}
 	if len(m.NodeId) != 32 {
 		return fmt.Errorf("invalid `nodeId` length; expected: 32, given: %v", len(m.NodeId))
 	}
-
 	if len(m.CommitmentAtxId) != 32 {
-		return fmt.Errorf("invalid `atxId` length; expected: 32, given: %v", len(m.CommitmentAtxId))
+		return fmt.Errorf("invalid `commitmentAtxId` length; expected: 32, given: %v", len(m.CommitmentAtxId))
 	}
 
 	numLabels := uint64(m.NumUnits) * uint64(m.LabelsPerUnit)
@@ -77,34 +79,68 @@ func Verify(p *shared.Proof, m *shared.ProofMetadata) error {
 		return fmt.Errorf("invalid indices set size; expected %d, given: %d", expectedSize, len(p.Indices))
 	}
 
-	difficulty := shared.ProvingDifficulty(numLabels, uint64(m.K1))
+	if options.verifyFunc == nil {
+		difficulty := shared.ProvingDifficulty2(numLabels, m.B, m.K1)
+		options.logger.Debug("verifying difficulty %d", difficulty)
+		options.verifyFunc = func(val uint64) bool {
+			return val < difficulty
+		}
+	}
+
 	buf := bytes.NewBuffer(p.Indices)
 	gsReader := shared.NewGranSpecificReader(buf, bitsPerIndex)
-	indicesSet := make(map[uint64]bool, m.K2)
+	indicesSet := make(map[uint64]struct{}, m.K2)
+
+	// create the ciphers for the specific nonce
+	d := shared.CalcD(numLabels, m.B)
+	offset := p.Nonce * uint32(d)
+	nonceBlock := uint8(offset / aes.BlockSize)
+
+	// since the value can be on a boundary between two blocks, we need to create two ciphers
+	ciphers := make([]cipher.Block, 2)
+	for i := uint8(0); i < 2; i++ {
+		c, err := oracle.CreateBlockCipher(m.Challenge, nonceBlock+i)
+		if err != nil {
+			return fmt.Errorf("creating cipher for block %d: %w", nonceBlock, err)
+		}
+		ciphers[i] = c
+	}
+
+	block := make([]byte, aes.BlockSize)
+	out := make([]byte, aes.BlockSize*2)
+	u64 := unsafe.Slice((*uint64)(unsafe.Pointer(&out[offset%aes.BlockSize])), 1)
+	mask := (uint64(1) << (d * 8)) - 1
 
 	for i := uint(0); i < uint(m.K2); i++ {
 		index, err := gsReader.ReadNextUintBE()
 		if err != nil {
 			return err
 		}
-		if indicesSet[index] {
+		if _, ok := indicesSet[index]; ok {
 			return fmt.Errorf("non-unique index: %d", index)
 		}
-		indicesSet[index] = true
+		indicesSet[index] = struct{}{}
 
-		res, err := WorkOracle(
+		// Recreate B-long labels block
+		labelStart := index * uint64(m.B)
+		labelEnd := labelStart + uint64(m.B) - 1
+		res, err := oracle.WorkOracle(
 			oracle.WithCommitment(oracle.CommitmentBytes(m.NodeId, m.CommitmentAtxId)),
-			oracle.WithPosition(index),
+			oracle.WithStartAndEndPosition(labelStart, labelEnd),
 			oracle.WithBitsPerLabel(uint32(m.BitsPerLabel)),
 		)
 		if err != nil {
 			return err
 		}
-		hash := FastOracle(m.Challenge, p.Nonce, res.Output)
-		hashNum := UInt64LE(hash[:])
-		if hashNum > difficulty {
-			return fmt.Errorf("fast oracle output is above the threshold; index: %d, label: %x, hash: %x, hashNum: %d, difficulty: %d",
-				index, res.Output, hash, hashNum, difficulty)
+		copy(block, res.Output)
+
+		ciphers[0].Encrypt(out[:aes.BlockSize], block)
+		ciphers[1].Encrypt(out[aes.BlockSize:], block)
+
+		val := u64[0] & mask
+		options.logger.Debug("verifying: index %d value %d", index, val)
+		if !options.verifyFunc(val) {
+			return fmt.Errorf("fast oracle output is doesn't pass difficulty check; index: %d, labels block: %x, value: %d", index, res.Output, val)
 		}
 	}
