+++ verifying/verifying_test.go
@@ -1,10 +1,10 @@
 package verifying
 
 import (
+	"bytes"
 	"context"
-	"encoding/binary"
 	"fmt"
-	"io"
+	"math/rand"
 	"testing"
 	"time"
 
@@ -12,22 +12,8 @@ import (
 
 	"github.com/spacemeshos/post/config"
 	"github.com/spacemeshos/post/initialization"
-	"github.com/spacemeshos/post/oracle"
-	"github.com/spacemeshos/post/persistence"
 	"github.com/spacemeshos/post/proving"
 	"github.com/spacemeshos/post/shared"
-)
-
-//lint:file-ignore SA1019 These are the tests for the deprecated functions and should still be executed as long as the deprecated functions haven't been removed.
-
-var (
-	nodeId          = make([]byte, 32)
-	commitmentAtxId = make([]byte, 32)
-	ch              = make(proving.Challenge, 32)
-
-	NewInitializer = initialization.NewInitializer
-	NewProver      = proving.NewProver
-	CPUProviderID  = initialization.CPUProviderID
 )
 
 func getTestConfig(tb testing.TB) (config.Config, config.InitOpts) {
@@ -36,16 +22,86 @@ func getTestConfig(tb testing.TB) (config.Config, config.InitOpts) {
 	opts := config.DefaultInitOpts()
 	opts.DataDir = tb.TempDir()
 	opts.NumUnits = cfg.MinNumUnits
-	opts.ComputeProviderID = int(CPUProviderID())
+	opts.ComputeProviderID = int(initialization.CPUProviderID())
 
 	return cfg, opts
 }
 
-func TestVerify(t *testing.T) {
+type testLogger struct {
+	shared.Logger
+
+	tb testing.TB
+}
+
+func (l testLogger) Info(msg string, args ...any)  { l.tb.Logf("\tINFO\t"+msg, args...) }
+func (l testLogger) Debug(msg string, args ...any) { l.tb.Logf("\tDEBUG\t"+msg, args...) }
+func (l testLogger) Error(msg string, args ...any) { l.tb.Logf("\tERROR\t"+msg, args...) }
+
+func BenchmarkVerifying(b *testing.B) {
+	for _, mB := range []uint32{8, 16} {
+		for _, k2 := range []uint32{170, 288, 500, 800} {
+			testName := fmt.Sprintf("256GiB/B=%d/k2=%d", mB, k2)
+
+			b.Run(testName, func(b *testing.B) {
+				benchmarkVerifying(b, mB, k2)
+			})
+		}
+	}
+}
+
+func benchmarkVerifying(b *testing.B, mB, k2 uint32) {
+	nodeId := make([]byte, 32)
+	commitmentAtxId := make([]byte, 32)
+
+	cfg, opts := getTestConfig(b)
+
+	m := &shared.ProofMetadata{
+		NodeId:          nodeId,
+		CommitmentAtxId: commitmentAtxId,
+		Challenge:       []byte("hello world, challenge me!!!!!!!"),
+		NumUnits:        opts.NumUnits,
+		BitsPerLabel:    cfg.BitsPerLabel,
+		LabelsPerUnit:   256 * 1024 * 1024 * 1024, // 256GiB
+		K1:              cfg.K1,
+		K2:              k2,
+		B:               mB,
+		N:               cfg.N,
+	}
+
+	numLabels := uint64(m.NumUnits) * uint64(m.LabelsPerUnit)
+	bitsPerIndex := uint(shared.BinaryRepresentationMinBits(numLabels))
+
+	var buf bytes.Buffer
+	gsWriter := shared.NewGranSpecificWriter(&buf, bitsPerIndex)
+	for i := uint32(0); i < m.K2; i++ {
+		require.NoError(b, gsWriter.WriteUintBE(uint64(i)))
+	}
+	require.NoError(b, gsWriter.Flush())
+
+	p := &shared.Proof{
+		Nonce:   rand.Uint32(),
+		Indices: buf.Bytes(),
+	}
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		start := time.Now()
+		err := Verify(p, m, withVerifyFunc(func(val uint64) bool { return true }))
+		require.NoError(b, err)
+		b.ReportMetric(time.Since(start).Seconds(), "sec/proof")
+	}
+}
+
+func Test_Verify(t *testing.T) {
 	r := require.New(t)
+	log := testLogger{tb: t}
+
+	nodeId := make([]byte, 32)
+	commitmentAtxId := make([]byte, 32)
+	ch := make(shared.Challenge, 32)
 
 	cfg, opts := getTestConfig(t)
-	init, err := NewInitializer(
+	init, err := initialization.NewInitializer(
 		initialization.WithNodeId(nodeId),
 		initialization.WithCommitmentAtxId(commitmentAtxId),
 		initialization.WithConfig(cfg),
@@ -54,109 +110,10 @@ func TestVerify(t *testing.T) {
 	r.NoError(err)
 	r.NoError(init.Initialize(context.Background()))
 
-	p, err := NewProver(cfg, opts.DataDir, nodeId, commitmentAtxId)
-	r.NoError(err)
-	proof, proofMetadata, err := p.GenerateProof(ch)
+	proof, proofMetadata, err := proving.Generate(context.Background(), ch, cfg, log, proving.WithDataSource(cfg, nodeId, commitmentAtxId, opts.DataDir))
 	r.NoError(err)
 
 	r.NoError(Verify(proof, proofMetadata))
-}
-
-func TestVerifyPow(t *testing.T) {
-	r := require.New(t)
-
-	cfg, opts := getTestConfig(t)
-	init, err := NewInitializer(
-		initialization.WithNodeId(nodeId),
-		initialization.WithCommitmentAtxId(commitmentAtxId),
-		initialization.WithConfig(cfg),
-		initialization.WithInitOpts(opts),
-	)
-	r.NoError(err)
-	r.NoError(init.Initialize(context.Background()))
-
-	m := &shared.VRFNonceMetadata{
-		NodeId:          nodeId,
-		CommitmentAtxId: commitmentAtxId,
-		NumUnits:        opts.NumUnits,
-		BitsPerLabel:    cfg.BitsPerLabel,
-		LabelsPerUnit:   uint64(opts.NumUnits) * cfg.LabelsPerUnit,
-	}
-	r.NoError(VerifyVRFNonce(init.Nonce(), m))
-}
-
-// TestLabelsCorrectness tests, for variation of label sizes, the correctness of
-// reading labels from disk (written in multiple files) when compared to a single label compute.
-// It is covers the following components: labels compute lib (package: oracle), labels writer (package: persistence),
-// labels reader (package: persistence), and the granularity-specific reader (package: shared).
-// it proceeds as follows:
-//  1. Compute labels, in batches, and write them into multiple files (prover).
-//  2. Read the sequence of labels from the files according to the specified label size (prover),
-//     and ensure that each one equals a single label compute (verifier).
-func TestLabelsCorrectness(t *testing.T) {
-	req := require.New(t)
-	if testing.Short() {
-		t.Skip("long test")
-	}
-
-	numFiles := 2
-	numFileBatches := 2
-	batchSize := 256
-	datadir := t.TempDir()
-
-	for bitsPerLabel := uint32(config.MinBitsPerLabel); bitsPerLabel <= config.MaxBitsPerLabel; bitsPerLabel++ {
-		t.Logf("bitsPerLabel: %v\n", bitsPerLabel)
-
-		// Write.
-		for i := 0; i < numFiles; i++ {
-			writer, err := persistence.NewLabelsWriter(datadir, i, uint(bitsPerLabel))
-			req.NoError(err)
-			for j := 0; j < numFileBatches; j++ {
-				numBatch := i*numFileBatches + j
-				startPosition := uint64(numBatch * batchSize)
-				endPosition := startPosition + uint64(batchSize) - 1
-
-				res, err := oracle.WorkOracle(
-					oracle.WithComputeProviderID(CPUProviderID()),
-					oracle.WithCommitment(oracle.CommitmentBytes(nodeId, commitmentAtxId)),
-					oracle.WithStartAndEndPosition(startPosition, endPosition),
-					oracle.WithBitsPerLabel(bitsPerLabel),
-				)
-				req.NoError(err)
-				req.NoError(writer.Write(res.Output))
-			}
-			_, err = writer.Close()
-			req.NoError(err)
-		}
-
-		// Read.
-		reader, err := persistence.NewLabelsReader(datadir, uint(bitsPerLabel))
-		req.NoError(err)
-		defer reader.Close()
-		gsReader := shared.NewGranSpecificReader(reader, uint(bitsPerLabel))
-		var position uint64
-		for {
-			label, err := gsReader.ReadNext()
-			if err != nil {
-				if err == io.EOF {
-					req.Equal(uint64(numFiles*numFileBatches*batchSize), position)
-					break
-				}
-				req.Fail(err.Error())
-			}
-
-			// Verify correctness.
-			labelCompute, err := oracle.WorkOracle(
-				oracle.WithCommitment(oracle.CommitmentBytes(nodeId, commitmentAtxId)),
-				oracle.WithPosition(position),
-				oracle.WithBitsPerLabel(bitsPerLabel),
-			)
-			req.NoError(err)
-			req.Equal(labelCompute, label, fmt.Sprintf("position: %v, bitsPerLabel: %v", position, bitsPerLabel))
-
-			position++
-		}
-	}
 }
 
 func Benchmark_Verify_Fastnet(b *testing.B) {
@@ -168,15 +125,17 @@ func Benchmark_Verify_Fastnet(b *testing.B) {
 
 	cfg, opts := getTestConfig(b)
 	cfg.BitsPerLabel = 8
-	cfg.K1 = 2000
+	cfg.K1 = 12
 	cfg.K2 = 4
-	cfg.LabelsPerUnit = 32
+	cfg.LabelsPerUnit = 32 // bytes
 	cfg.MaxNumUnits = 4
 	cfg.MinNumUnits = 2
+	cfg.N = 32
+	cfg.B = 2
 
 	opts.NumUnits = cfg.MinNumUnits
 
-	init, err := NewInitializer(
+	init, err := initialization.NewInitializer(
 		initialization.WithNodeId(nodeId),
 		initialization.WithCommitmentAtxId(commitmentAtxId),
 		initialization.WithConfig(cfg),
@@ -186,17 +145,13 @@ func Benchmark_Verify_Fastnet(b *testing.B) {
 	r.NoError(init.Initialize(context.Background()))
 
 	for i := 0; i < b.N; i++ {
-		binary.BigEndian.PutUint64(ch, uint64(opts.NumUnits))
-
-		p, err := NewProver(cfg, opts.DataDir, nodeId, commitmentAtxId)
-		r.NoError(err)
-
-		proof, proofMetadata, err := p.GenerateProof(ch)
+		rand.Read(ch)
+		proof, proofMetadata, err := proving.Generate(context.Background(), ch, cfg, &shared.DisabledLogger{}, proving.WithDataSource(cfg, nodeId, commitmentAtxId, opts.DataDir))
 		r.NoError(err)
 
 		b.StartTimer()
 		start := time.Now()
-		require.NoError(b, Verify(proof, proofMetadata))
+		r.NoError(Verify(proof, proofMetadata))
 		b.ReportMetric(time.Since(start).Seconds(), "sec/proof")
 		b.StopTimer()
 	}
